Remove duplicate helpers, extract non-duplicate helpers
