+++ iati/tests/home_functional_tests.py
@@ -1,9 +1,6 @@
 """A module of functional tests for the home page."""
-import os
 import pytest
-from django.utils.text import slugify
-from tests.base_functional_tests import click_obscured, view_live_page
-from iati.urls import ADMIN_SLUG
+from tests import helper_functions
 
 
 HOME_PAGE = {
@@ -11,12 +8,6 @@ HOME_PAGE = {
     'heading': 'Test International Aid Transparency Initiative',
     'excerpt': 'This is an excerpt for the Home page'
 }
-
-
-def publish_changes(admin_browser):
-    """Publish changes made in the CMS to the live page."""
-    click_obscured(admin_browser, admin_browser.find_by_xpath('//div[@class="dropdown-toggle icon icon-arrow-up"]').first)
-    click_obscured(admin_browser, admin_browser.find_by_text('Publish').first)
 
 
 @pytest.mark.django_db
@@ -37,8 +28,8 @@ class TestHomePage():
         # Fill the heading field
         admin_browser.fill('heading_en', HOME_PAGE['heading'])
         # Publish new page heading
-        publish_changes(admin_browser)
-        view_live_page(admin_browser, HOME_PAGE['title'])
+        helper_functions.publish_changes(admin_browser)
+        helper_functions.view_live_page(admin_browser, HOME_PAGE['title'])
         assert admin_browser.find_by_text(HOME_PAGE['heading'])
 
     def test_can_edit_home_page_excerpt(self, admin_browser):
@@ -49,8 +40,8 @@ class TestHomePage():
         # Fill the excerpt field
         admin_browser.fill('excerpt_en', HOME_PAGE['excerpt'])
         # Publish new page excerpt
-        publish_changes(admin_browser)
-        view_live_page(admin_browser, HOME_PAGE['title'])
+        helper_functions.publish_changes(admin_browser)
+        helper_functions.view_live_page(admin_browser, HOME_PAGE['title'])
         assert admin_browser.find_by_text(HOME_PAGE['excerpt'])
 
 
@@ -70,42 +61,6 @@ TERMS_PAGE = {
 }
 
 
-def enter_page_content(admin_browser, tab_name, cms_field, cms_content):
-    """Add title and slug to a page in the CMS.
-
-    Args:
-        tab_name (str): The name of a tab on an edit page of the CMS.
-        cms_field (str): The name of the field in the CMS you want to fill.
-        cms_content (str): The text content you want to fill the field with.
-
-    """
-    tab = admin_browser.find_by_text(tab_name)
-    _ = tab[0].__dict__['_element'].location_once_scrolled_into_view
-    tab[0].click()
-    elem = admin_browser.find_by_css("[name='{}']".format(cms_field))
-    _ = elem[0].__dict__['_element'].location_once_scrolled_into_view
-    admin_browser.fill(cms_field, cms_content)
-
-
-def create_standard_home_page(admin_browser, page_type, fixed_page_type, page_title):
-    """Create a child page in the CMS.
-
-    Args:
-        page_type (str): The verbose name of the page model type you want to click on.
-        fixed_page_type (str): The value of the fixed page dropdown (e.g. "privacy" for "PrivacyPage")
-        page_title (str): The title of the page you are editing.
-
-    """
-    admin_browser.visit(os.environ['LIVE_SERVER_URL'] + "/{}/pages/3/".format(ADMIN_SLUG))
-    admin_browser.find_by_text('Add child page').click()
-    admin_browser.find_by_text(page_type).click()
-    dropdown = admin_browser.find_by_id("id_fixed_page_type")[0]
-    dropdown.select(fixed_page_type)
-    enter_page_content(admin_browser, 'English', 'title_en', page_title)
-    enter_page_content(admin_browser, 'Promote', 'slug_en', slugify(page_title))
-    publish_changes(admin_browser)
-
-
 @pytest.mark.django_db
 class TestFixedStandardPages():
     """Tests to see if fixed standard pages are properly linked"""
@@ -115,8 +70,8 @@ class TestFixedStandardPages():
     ])
     def test_can_create_standard_pages(self, admin_browser, fixed_page):
         """Check that when an event index child page is created it appears in the website."""
-        create_standard_home_page(admin_browser, fixed_page['page_type'], fixed_page['fixed_page_type'], fixed_page['title'])
-        view_live_page(admin_browser, fixed_page['title'])
+        helper_functions.create_standard_home_page(admin_browser, fixed_page['page_type'], fixed_page['fixed_page_type'], fixed_page['title'])
+        helper_functions.view_live_page(admin_browser, fixed_page['title'])
         assert admin_browser.title == fixed_page['title']
         admin_browser.click_link_by_text(fixed_page['base_link_text'])
         assert admin_browser.title == fixed_page['title']
